ipa: rpi: Add an HDR algorithm to drive multi-channel AGC
This implements a simple HDR algorithm that is sufficient for the vc4
platform, where images are not combined in any way. Instead, it
allowsan application to drive the AGC in a multi-channel HDR type of
mode, such as to produce short and long exposure images.

On the PiSP platform, certain HDR blocks are programmed with dummy
values until HDR has been implemented properly for the platform.

Sufficient plumbing is added to enable the libcamera HDR controls and
metadata to work.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Reviewed-by: Naushir Patuck <naush@raspberrypi.com>