ipa: rpi: Upgrade to HDR to support tonemapping
The HDR algorithm is extended so that it supports controlling the
tonemapping block on platforms that have it. The tonemapping curve may
be set explicitly or calculated dynamically.

The tonemapping curve is reset whenever the HDR mode changes. The IPA
base code ensures that the HDR algorithm runs, in order to update
these curves if necessary, whenever it sees the HDR mode change.

A subsequent commit will add control of the stitching functionality.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>