ipa: rpi: Support dual exposure stitching in HDR
Adds support for stitching together long and short exposure images.

We take care to enable the stitch output block when a multi-exposure
HDR mode of this type starts, and the stitch itself once we have both
long and short frames. We signal the pipeline handler that it needs to
do a swap of the two stitch buffers when the channel (long or short)
changes.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>