ipa: rpi: pisp: Update HDR tuning
The latest HDR tuning is applied to all sensors. The changes are
identical for all sensors.

The principle changes are to allow a few more pixels to saturate, and
to bring up the very darkest areas slightly more, creating a more even
handed effect in most situations.

Exposure modes are also updated so that they will work as expected at
30fps, and you can switch between normal/short and long if you want
longer exposure captures.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>